requirements.txt
asgiref==3.7.2
Django==4.2.9
djangorestframework==3.14.0
pytz==2023.4
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.9
djangorestframework==3.14.0
pytz==2023.4
sqlparse==0.4.4
isort==5.13.2
black==24.1.1